feat: deleting statuses
This commit adds support for deleting statuses.

Before sending the delete request to the instance, Enbas will first
verify that the status that the user wants to delete actually belongs to
them.

The user has the option to save the text of the deleted status. This
will be written to a text file within the cache directory.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/48
Resolves: https://codeflow.dananglin.me.uk/apollo/enbas/issues/44