feat: switch to client-server architecture
This commit updates the architecture of Enbas by switching to the
client-server architecture. The motivation for such change is to remove
the need to read the credentials file and create the GTSClient value
every time the user issues a command to Enbas, and to also support
future features such as adding in-memory cache to the server side for
paging through timelines, notifications and other lists.

The app now introduces server mode that launches an RPC server. For now
we are using the net/rpc package from the standard library. The server
registers the qualifying methods from the value of the GTSClient type
and listens for connections using a unix socket file that the user
must specify in their configuration file. The relevant executor in the
executors package is now a client to the RPC server where it connects
and sends requests to the server. The server processes the client's
request and updates the value of the output that the client passes to
the server (by reference).

The user can launch the server with the new server executor. They can
set a timeout that shuts down the sever after a certain amount of idle
time, or they can run the server without the timeout using the
'--no-idle-timeout' flag which allows the server to run indefinitely. If
the user issues a command to Enbas and the server is not running, the
executor with automatically launch the server using the settings in the
configuration file before processing the user's request.

Other notable changes:

- fix: When displaying an edited list, retrieve and show the accounts
  that are a part of that list.
- fix(BREAKING): For creating statuses, replace the enable-federation
  flag with local-only as LocalOnly replaces EnableFederation in
  GoToSocial.
- feat: add the list-exclusive flag for creating lists.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/62