refactor: update code for running the pager
Update the code for running the pager so that the logic is clearer.
First, we make sure that the first word in the user configured pager
config is an executable, then we write the text data to the pipe, then
we connect the pipe to the pager and run the pager to display the text
on screen.

Additional changes:

The Printer's formatDate and formatDateTime methods have been converted
to named functions. They don't need to be methods of the Printer type.