refactor: replace the Printer type in printer pkg
This is the second part of the refactoring of the printer package. In
this commit the Printer type is removed and its methods have been
converted to named functions. A new Settings type is introduced and most
of the converted functions return functions after injecting the user
defined settings from Settings.

The PrintVersion function is also updated so that it now uses a Go
template to print the application's build information to standard out.