refactor: get user's account ID from gtsclient
Whenever an executor needs to get the user's account ID they will now
get it directly from the gtsclient package. This is in preparation for
the in-memory cache that will be added in later.

When the cache is added, the RPC server can store the ID there to
reduce the amount of requests made to the GTS instance.