feat(BREAKING): CLI redesign
This commit redesigns the CLI to make it much more easier to use.
The redesign now follows two simpler and strict formats:
  enbas [top level flags] action target [flags]
  enbas [top level flags] action target-1 <preposition> target-2
  <flags>.
For example, to show a status the previous command was:
  enbas show --type status --status-id <STATUS_ID>
Now the command to do this is:
  enbas show status --status-id <STATUS_ID>

Users will be able to simplify this format even further with the use of
aliases which will be introduced in the near future.

The CLI parser has been upgraded to parse the command using the
formatting rules above and it will validate it before the executor
performs the desired operation.

A new source of truth for the CLI is created in
'definitions/definitions.json' which replaces the old
'schema/enbas_cli_schema.json' file. For now this file is used to
generate the code in 'internal/cli' and the targetFuncMap map in
'internal/executor'.

This commit also adds new functionality for displaying the help
documentation. This is now generated using go templates and the internal
printer package. The help documentation now shows exactly which flags
are associated with which type of operation (e.g. the flags that can be
used for 'create status').

The help functionality still needs a lot of work but it is an
improvement on the previous implementation.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/69
Resolves: https://codeflow.dananglin.me.uk/apollo/enbas/issues/63