docs: generating the Enbas manual
This is the first iteration of the Enbas manual.

For now the operational guide (enbas.1) and the configuration guide
(enbas.5) is generated. The manual is generated with the new
./cmd/docgen code and can be executed via mage
(mage build:documentation).

The manual is generated using the templates within
./cmd/docgen/templates/manpages and the source of truth. The manpages
are generated are generated under __build/share/man.

The SoT has been updated by adding a description for each 'TargetAction'
or 'TargetOnTargetAction'. These descriptions are displayed on the
operations manual and may also be used in the usage documentation in the
future.

An example configuration file is generated alongside the manpages. It is
generated to __build/share/doc/enbas/examples/config.json.

The mage target for building the binary is updated. It is now under the
new Build namespace and is called build:binary. The binary is now built
to __build/bin/enbas by default.

The install target is removed. This is to give the user (or package
manager) more control on how they'd like to install the application on
their system. The documentation will need to be updated to reflect this
(which will be handled in another commit).

The documents docs/03_configuration.md and docs/04_user_manual.md are
removed in favour of the new manpages. The remaining documentation is
moved to doc/ until they are also removed in the future.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/77