fix: server mode persistence using client sessions
Ensure that the server mode remains active if one or more client
sessions are in progress even if the idle timeout has passed.

This commit adds a session store and each client session generates a
session ID and adds it to the session store. Once the client session has
finished the session ID is removed from the store.

The server remains active until that are no more session IDs in the
session store.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/85