fix: update the usage target
Update the usage target by improving the templates and ensuring that all
operations (including target to target operations) are documented and
available to the user.

The action flag is replaced with the operation flag so that users can
see the usage documentation for an operation such as 'create status'.

The README.md file is updated to by including instructions on how users
can use the usage target to view the usage documentation.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/91