fix: add a new Auth type for the GTSClient value
Add a new Auth type for the GTSClient value to ensure that the value's
concurrent reads and writes to/from its own authentication details can
be performed safely.

The user's account ID is stored in the new Auth type to reduce the
number of API requests made to the GtS instance.

PR: https://codeflow.dananglin.me.uk/apollo/enbas/pulls/92